Latest changes

+++ src/yapca.sh
@@ -19,27 +19,44 @@ set -exo pipefail
 # Argbash is a bash code generator used to get arguments parsing right.
 # Argbash is FREE SOFTWARE, see https://argbash.io for more info
 # Generated online by https://argbash.io/generate
+# ./test/bats/bin/bats test/yapca_basic.bats
+# ./test/bats/bin/bats test/yapca_init.bats
+# ./test/bats/bin/bats test/yapca_encini.bats
 
 SCRIPTNAME="${0##*/}"
 SCRIPTNAME="${SCRIPTNAME%%.*}"
+readonly SCRIPTNAME
 SCRIPTPATH=$(dirname "$0")
 SCRIPTPATH="$(cd "${SCRIPTPATH}" && pwd -P)/"
+readonly SCRIPTPATH
 SCRIPTUMASK="$(umask)"
+readonly SCRIPTUMASK
 CURRENTDIR="$(pwd -P)/"
+readonly CURRENTDIR
 
 # die codes
-ERR_ARGBASH=101
-ERR_ENCRYPT_INI=102
-ERR_READ_INI=103
-ERR_UNKNOWN_COMMAND=104
+readonly ERR_ARGBASH=101
+readonly ERR_ENCRYPT_INI=102
+readonly ERR_READ_INI=103
+readonly ERR_UNKNOWN_COMMAND=104
 
 # cert related constants
-SERVER_KEYLENGTH=2048
+readonly SERVER_KEYLENGTH=2048
+
+# Define trace levels
+readonly TRACE_LEVEL_ERROR=1
+readonly TRACE_LEVEL_WARN=2
+readonly TRACE_LEVEL_INFO=3
+readonly TRACE_LEVEL_DEBUG=4
+
+# Default trace level
+readonly DEFAULT_TRACE_LEVEL=${TRACE_LEVEL_INFO}
 
 die() {
   local _ret="${2:-1}"
   test "${_PRINT_HELP:-no}" = yes && print_help >&2
   echo "$1" >&2
+	yapca_trace "${TRACE_LEVEL_ERROR}" "End fail"
   umask "${SCRIPTUMASK}"
   exit "${_ret}"
 }
@@ -170,12 +187,112 @@ assign_positional_args() {
   done
 }
 
+# Define function to print trace messages
+yapca_trace() {
+  local level=${1}
+  local message=${2}
+
+  # Check if trace level is greater than or equal to the threshold
+  if [[ ${level} -le ${DEFAULT_TRACE_LEVEL} ]]; then
+    case ${level} in
+      "${TRACE_LEVEL_ERROR}") echo "[ERROR]: ${message}";;
+      "${TRACE_LEVEL_WARN}") echo "[WARNING]: ${message}";;
+      "${TRACE_LEVEL_INFO}") echo "[INFO]: ${message}";;
+      "${TRACE_LEVEL_DEBUG}") echo "[DEBUG]: ${message}";;
+    esac
+  fi
+}
+
 function yapca_random_string() {
   # ensure $1 is a number
   local l_hex_digits
-	l_hex_digits="$(printf '%d' "${1}")"
+	printf -v l_hex_digits '%d' "${1}"
   # generate random hex string
   printf '%8s' "$(od -An -N"${l_hex_digits}" -tx1 /dev/urandom | tr -d '[:blank:]')"
+}
+
+function yapca_encrypt_string() {
+	# processed string
+	local p_str_dec
+	local p_str_enc
+	# salt
+	local p_salt_front
+	local p_salt_back
+	# salt length
+	local p_len_front
+	local p_len_back
+	# encrypt parameters
+	p_enc_cmd=(
+		"enc"
+		"-aes-256-cbc"
+		"-pbkdf2"
+		"-base64"
+		"-A"
+		"-pass"
+		"file:${2}"
+	)
+	
+	p_str_dec="${1}"
+	# Generate a random prefix and postfix salt
+	p_salt_front="$("${_arg_openssl}" rand $((RANDOM % 16 + 16)) | tr -dc '[:alnum:]' )"
+	p_salt_back="$("${_arg_openssl}" rand $((RANDOM % 16 + 16)) | tr -dc '[:alnum:]' )"
+	# get the lengths
+	p_len_front=${#p_salt_front}
+	p_len_back=${#p_salt_back}
+	# encrypt
+	# pattern "<prefix_len><prefix><password><postfix><postfix_len>"
+	# <prefix_len> is always 2 chars
+	# <postfix_len> is always 2 chars
+	printf -v p_str_enc "%02x%s%s%s%02x" "${p_len_front}" "${p_salt_front}" "${p_str_dec}" "${p_salt_back}" "${p_len_back}"
+	p_str_enc="$(echo -e "${p_str_enc}" | "${_arg_openssl}" "${p_enc_cmd[@]}")"
+	printf "%s" "${p_str_enc}"
+}
+
+function yapca_decrypt_string() {
+	# processed string
+	local p_str_enc
+	local p_str_dec
+	local p_str_len
+	# salt length
+	local p_len_afront
+	local p_len_back
+	# temporary strings
+	local p_tmp_front
+	local p_tmp_back
+	local p_tmp_dec
+	# decrypt parameters
+	p_dec_cmd=(
+		"enc"
+		"-aes-256-cbc"
+		"-pbkdf2"
+		"-base64"
+		"-d"
+		"-pass"
+		"file:${2}"
+	)
+	
+	p_str_enc="${1}"
+	# decrypt
+	# pattern "<prefix_len><prefix><password><postfix><postfix_len>"
+	# <prefix_len> is always 2 chars
+	# <postfix_len> is always 2 chars
+	# decoded string
+	p_str_dec="$(echo -e "${p_str_enc}" | "${_arg_openssl}" "${p_dec_cmd[@]}")"
+	# extract pass
+	# determine the length of the prefix salt
+	printf -v p_len_front "%d" "0x${p_str_dec:0:2}"
+	# remove the front salt ^<frontsaltlength><frontsalt>
+	p_tmp_front="${p_str_dec:0:2+${p_len_front}}"
+	p_tmp_dec=${p_str_dec#"${p_tmp_front}"}
+	# determine the back salt <backsalt><backsaltlength>$
+	p_tmp_front="${p_tmp_dec%??}"
+	p_tmp_back=${p_tmp_dec#"${p_tmp_front}"}
+	printf -v p_len_back "%d" "0x${p_tmp_back}"
+	# determine the password length
+	p_str_len=${#p_tmp_dec}-2-${p_len_back}
+	# decrypted string withoud salt
+	p_str_dec=${p_str_dec:2+${p_len_front}:${p_str_len}}
+	printf "%s" "${p_str_dec}"
 }
 
 ##
@@ -461,10 +578,23 @@ function yapca_server_csr() {
 }
 
 ##
+## Create a certificate by signing the request
+##
+function yapca_server_crt() {
+
+	openssl ca -config intermediate/openssl.cnf \
+      -extensions server_cert -days 375 -notext -md sha256 \
+      -in intermediate/csr/www.example.com.csr.pem \
+      -out intermediate/certs/www.example.com.cert.pem
+	chmod 444 intermediate/certs/www.example.com.cert.pem
+}
+
+##
 ## Decrypt password
 ##
 function yapca_decrypt_pass() {
   l_enc_str="${1}"
+	l_dec_str="${l_enc_str}"
 	l_name_base=".${SCRIPTNAME}.masterpass"
 	l_name_dir="${_ca_dir}/"
 	# Check if password file name is in env
@@ -476,26 +606,14 @@ function yapca_decrypt_pass() {
 	# check if the password is encrypted
 	if [[ "${l_enc_str}" == txt::* ]]; then
 		# remove prefix
-		l_enc_str=${l_enc_str#?????}
+		l_dec_str="${l_enc_str#?????}"
 	# check if we have a master password
 	elif [[ "$(find "${l_name_dir}." ! -name . -prune -type f -name "${l_name_base}" -perm 400)" == "${l_name_dir}./${l_name_base}" && "${l_enc_str}" == enc::* ]]; then
-		a_cmd=(
-			"enc"
-			"-aes-256-cbc"
-			"-pbkdf2"
-			"-base64"
-			"-d"
-			"-pass"
-			"file:${l_name_dir}${l_name_base}"
-		)
 		# remove prefix
-		l_enc_str=${l_enc_str#?????}
-		l_enc_str="$(echo -e "${l_enc_str}" | "${_arg_openssl}" "${a_cmd[@]}")"
-		# remove salt
-		l_enc_str=${l_enc_str#????????????????}
-		l_enc_str=${l_enc_str%????????}
+		l_enc_str="${l_enc_str#?????}"
+		l_dec_str="$(yapca_decrypt_string "${l_enc_str}" "${l_name_dir}${l_name_base}")"
   fi
-  echo "${l_enc_str}"
+  echo "${l_dec_str}"
 }
 
 ##
@@ -512,46 +630,35 @@ function yapca_encrypt_pass() {
 		l_name_dir="$(cd "${l_name_dir}" && pwd -P)/"
 	fi
 	if [[ "$(find "${l_name_dir}." ! -name . -prune -type f -name "${l_name_base}" -perm 400)" == "${l_name_dir}./${l_name_base}" ]]; then
-    a_cmd=(
-      "enc"
-      "-aes-256-cbc"
-      "-pbkdf2"
-      "-base64"
-      "-A"
-      "-pass"
-			"file:${l_name_dir}${l_name_base}"
-    )
     # write to ini file
-    if [[ -n ${l_ini} && -f ${l_ini} ]]; then
-      if [[ -w ${l_ini} ]]; then
-        # rewrite to temporary file with encrypted passwords
-        echo -n >"${_arg_tmp}/encrypted.ini"
-        while IFS='' read -r INILINE || [[ -n "${INILINE}" ]]; do
-          {
-            if [[ "${INILINE}" == pass* ]]; then
-              l_name="${INILINE%%=*}"
-              l_pass="${INILINE#*=}"
-              if [[ "${l_pass}" == txt::* ]]; then
-                # encode with salt
-                l_salt_front="$(yapca_random_string 8)"
-                l_salt_back="$(yapca_random_string 4)"
-                l_pass=${l_pass#?????}
-                l_pass="$(echo -e "${l_salt_front}${l_pass}${l_salt_back}" | "${_arg_openssl}" "${a_cmd[@]}")"
-                printf '%s=enc::%s\n' "${l_name}" "${l_pass//=/\\075}"
-              else
-                printf '%s\n' "${INILINE}"
-              fi
-            else
-              printf '%s\n' "${INILINE}"
-            fi
-          } >>"${_arg_tmp}/encrypted.ini"
-        done <"${l_ini}"
-        # replace plaintext password ini with the encrypted one
-        mv "${l_ini}" "${l_ini%.*}.bak" && cp "${_arg_tmp}/encrypted.ini" "${l_ini}"
-      else
-        die "File: ${l_ini} is not writable." "${ERR_ENCRYPT_INI}"
-      fi
-    fi
+		if [[ -n ${l_ini:-} && -f ${l_ini:-} && -w ${l_ini:-} ]]; then
+			# rewrite to temporary file with encrypted passwords
+			echo -n >"${_arg_tmp}/encrypted.ini"
+			while IFS='' read -r INILINE || [[ -n "${INILINE}" ]]; do
+				{
+					if [[ "${INILINE}" == pass* ]]; then
+						l_name="${INILINE%%=*}"
+						l_pass="${INILINE#*=}"
+						if [[ "${l_pass}" == txt::* ]]; then
+							# remove the txt:: prefix
+							l_pass="${l_pass#?????}"
+							# encrypt password
+							l_enc_pass="$(yapca_encrypt_string "${l_pass}" "${l_name_dir}${l_name_base}")"
+							# write back
+							printf '%s=enc::%s\n' "${l_name}" "${l_enc_pass//=/\\075}"
+						else
+							printf '%s\n' "${INILINE}"
+						fi
+					else
+						printf '%s\n' "${INILINE}"
+					fi
+				} >>"${_arg_tmp}/encrypted.ini"
+			done <"${l_ini}"
+			# replace plaintext password ini with the encrypted one
+			mv "${l_ini}" "${l_ini%.*}.bak" && cp "${_arg_tmp}/encrypted.ini" "${l_ini}"
+		else
+			die "File: ${l_ini} is not writable." "${ERR_ENCRYPT_INI}"
+		fi
   else
     die "Masterpassword file: ${l_name_dir}${l_name_base} does not exist or has wrong permissions." "${ERR_ENCRYPT_INI}"
   fi
@@ -592,25 +699,30 @@ fi
 # 10) name "default.ini " + script directory
 l_inifiles=()
 if [[ -n ${_arg_ini} ]]; then
-  l_inifiles+=("${_arg_ini}")
-  l_inifiles+=("${_arg_ini}.ini")
-  l_inifiles+=("${CURRENTDIR}${_arg_ini}.ini")
-  l_inifiles+=("${CURRENTDIR}ini/${_arg_ini}.ini")
-  l_inifiles+=("${SCRIPTPATH}${_arg_ini}.ini")
-  l_inifiles+=("${SCRIPTPATH}ini/${_arg_ini}.ini")
-  l_inifiles+=("${CURRENTDIR}${_arg_ini}")
-  l_inifiles+=("${CURRENTDIR}ini/${_arg_ini}")
-  l_inifiles+=("${SCRIPTPATH}${_arg_ini}")
-  l_inifiles+=("${SCRIPTPATH}ini/${_arg_ini}")
+  l_inifiles+=(
+    "${_arg_ini}"
+    "${_arg_ini}.ini"
+    "${CURRENTDIR}${_arg_ini}.ini"
+    "${CURRENTDIR}ini/${_arg_ini}.ini"
+    "${SCRIPTPATH}${_arg_ini}.ini"
+    "${SCRIPTPATH}ini/${_arg_ini}.ini"
+    "${CURRENTDIR}${_arg_ini}"
+    "${CURRENTDIR}ini/${_arg_ini}"
+    "${SCRIPTPATH}${_arg_ini}"
+    "${SCRIPTPATH}ini/${_arg_ini}"
+  )
 fi
-l_inifiles+=("${CURRENTDIR}${SCRIPTNAME}.ini")
-l_inifiles+=("${CURRENTDIR}ini/${SCRIPTNAME}.ini")
-l_inifiles+=("${SCRIPTPATH}${SCRIPTNAME}.ini")
-l_inifiles+=("${SCRIPTPATH}ini/${SCRIPTNAME}.ini")
-l_inifiles+=("${CURRENTDIR}default.ini")
-l_inifiles+=("${CURRENTDIR}ini/default.ini")
-l_inifiles+=("${SCRIPTPATH}default.ini")
-l_inifiles+=("${SCRIPTPATH}ini/default.ini")
+l_inifiles+=(
+  "${CURRENTDIR}${SCRIPTNAME}.ini"
+  "${CURRENTDIR}ini/${SCRIPTNAME}.ini"
+  "${SCRIPTPATH}${SCRIPTNAME}.ini"
+  "${SCRIPTPATH}ini/${SCRIPTNAME}.ini"
+  "${CURRENTDIR}default.ini"
+  "${CURRENTDIR}ini/default.ini"
+  "${SCRIPTPATH}default.ini"
+  "${SCRIPTPATH}ini/default.ini"
+)
+
 for ((i = 0; i < ${#l_inifiles[@]}; i++)); do
   if [[ -f ${l_inifiles[$i]} && -r ${l_inifiles[$i]} ]]; then
     _arg_ini="${l_inifiles[$i]}"
@@ -621,34 +733,32 @@ done
 ################################################################################
 # read ini file if defined
 # do not overwrite already existing parameters from command line
-if [[ -n ${_arg_ini} && -f ${_arg_ini} ]]; then
-  if [[ -r ${_arg_ini} ]]; then
-    while IFS='=' read -r INIVAR INIVAL; do
-      if [[ ${INIVAR} == [* ]]; then
-        INIVAR="${INIVAR#?}"
-        INISEC="${INIVAR%%?}"
-				# this part reads only the ca and intermediate sections from ini
-				if [[ "${INISEC}" != "ca" && "${INISEC}" != "intermediate" ]]; then
-					break
-				fi
-      elif [[ -n ${INIVAL} ]]; then
-        ARGVAR="_${INISEC}_${INIVAR}"
-        eval "ARGVAL=\$$ARGVAR"
-        if [[ -z ${ARGVAL} ]]; then
-          if [[ ${ARGVAR} == "_arg_parameter" ]]; then
-            for _last_positional in ${INIVAL}; do
-              _positionals+=("$_last_positional")
-              _positionals_count=$((_positionals_count + 1))
-            done
-          else
-            declare "_${INISEC}_${INIVAR}=${INIVAL}"
-          fi
+if [[ -n ${_arg_ini:-} && -f ${_arg_ini:-} && -r ${_arg_ini:-} ]]; then
+  while IFS='=' read -r INIVAR INIVAL; do
+    if [[ ${INIVAR} == [* ]]; then
+      INIVAR="${INIVAR#?}"
+      INISEC="${INIVAR%%?}"
+			# this part reads only the ca and intermediate sections from ini
+			if [[ "${INISEC}" != "ca" && "${INISEC}" != "intermediate" ]]; then
+				break
+			fi
+    elif [[ ${INIVAL:-} ]]; then
+      ARGVAR="_${INISEC}_${INIVAR}"
+      ARGVAL=${!ARGVAR:-}
+      if [[ -z ${ARGVAL} ]]; then
+        if [[ ${ARGVAR} == "_arg_parameter" ]]; then
+          for _last_positional in ${INIVAL}; do
+            _positionals+=("$_last_positional")
+            _positionals_count=$((_positionals_count + 1))
+          done
+        else
+          declare "_${INISEC}_${INIVAR}=${INIVAL}"
         fi
       fi
-    done <"${_arg_ini}"
-  else
-    die "File: ${_arg_ini} is not readable." "${ERR_READ_INI}"
-  fi
+    fi
+  done < "${_arg_ini}"
+else
+  die "File: ${_arg_ini:-} is not readable." "${ERR_READ_INI}"
 fi
 
 # set defaults here if not read from ini file
@@ -686,22 +796,43 @@ _intermediate_passout="$(yapca_decrypt_pass "${_intermediate_passout}")"
 _intermediate_passcsr="$(yapca_decrypt_pass "${_intermediate_passcsr}")"
 
 umask 0077
+yapca_trace "${TRACE_LEVEL_INFO}" "Begin"
 case "${_arg_positional}" in
 init)
+	yapca_trace "${TRACE_LEVEL_INFO}" "Initialize CA"
+	# echo "password" > .yapca.masterpass
+	# mkdir -p CA; cp .yapca.masterpass CA/
+	# ./src/yapca.sh -p /home/yapca/CA/ init
   yapca_initialize_ca
   ;;
 encini)
+	yapca_trace "${TRACE_LEVEL_INFO}" "Encode passwords in the INI file"
+	# ./src/yapca.sh encini
   yapca_encrypt_pass "${_arg_ini}"
   ;;
+decini)
+	yapca_trace "${TRACE_LEVEL_INFO}" "Print decrypted passwords from the INI file"
+	# ./src/yapca.sh decini
+	printf "CA::passin \t\"%s\"\n" "${_ca_passin}"
+	printf "CA::passout\t\"%s\"\n" "${_ca_passout}"
+	printf "CA::passcsr\t\"%s\"\n" "${_ca_passcsr}"
+	printf "intermediate::passin \t\"%s\"\n" "${_intermediate_passin}"
+	printf "intermediate::passout\t\"%s\"\n" "${_intermediate_passout}"
+	printf "intermediate::passcsr\t\"%s\"\n" "${_intermediate_passcsr}"
+	;;
 server)
+	yapca_trace "${TRACE_LEVEL_INFO}" "Creating a new server key"
 	yapca_server_key "keyname" 2048
 	yapca_server_csr "keyname"
 	;;
 noop)
+	yapca_trace "${TRACE_LEVEL_INFO}" "No operation performed"
 	;;
 *)
+	yapca_trace "${TRACE_LEVEL_ERROR}" "Unknown command"
   die "FATAL ERROR: Unknown command" ${ERR_UNKNOWN_COMMAND}
   ;;
 esac
+yapca_trace "${TRACE_LEVEL_INFO}" "End success"
 umask "${SCRIPTUMASK}"
 exit 0
